+++ package.json
@@ -6,6 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
+    "test": "jest",
     "server": "node -r dotenv/config server/index.js dotenv_config_path=server/.env",
     "lint-next": "next lint",
     "lint": "eslint --fix --c .eslintrc.json --ext .js,.jsx,.mjs .; prettier --write --config ./.prettierrc .",
@@ -35,9 +36,11 @@
     "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "jest": "^29.7.0",
     "nodemon": "^3.0.3",
     "postcss": "^8",
     "prettier": "^3.2.4",
+    "supertest": "^6.3.4",
     "tailwindcss": "^3.3.0"
   }
 }
